# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AuditIQ API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /engagements:
    get:
      operationId: listEngagements
      tags: [engagements]
      summary: List engagements
      responses:
        "200":
          description: List of engagements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Engagement"
    post:
      operationId: createEngagement
      tags: [engagements]
      summary: Create engagement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEngagementBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Engagement"
  /engagements/{id}:
    get:
      operationId: getEngagement
      tags: [engagements]
      summary: Get engagement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Engagement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Engagement"
    delete:
      operationId: deleteEngagement
      tags: [engagements]
      summary: Delete engagement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Journal Entries (Upload)
...
  /engagements/{id}/dashboard:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats for an engagement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /engagements/{id}/heatmap/users:
    get:
      operationId: getUserHeatmap
      tags: [dashboard]
      summary: Get risk heatmap by user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User heatmap data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserHeatmapRow"
  /engagements/{id}/heatmap/time:
    get:
      operationId: getTimeHeatmap
      tags: [dashboard]
      summary: Get risk heatmap by posting time
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Time heatmap data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeHeatmapRow"
  /engagements/{id}/risk-distribution:
    get:
      operationId: getRiskDistribution
      tags: [dashboard]
      summary: Get risk distribution breakdown
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Risk distribution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskDistribution"
  /engagements/{id}/benford:
    get:
      operationId: getBenfordAnalysis
      tags: [dashboard]
      summary: Get Benford's Law analysis for an engagement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Benford analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BenfordAnalysis"
  /engagements/{id}/duplicates:
    get:
      operationId: getDuplicates
      tags: [dashboard]
      summary: Get potential duplicate entries
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Duplicate groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DuplicateGroup"

  # Reports
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Auth
...
    Engagement:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        clientName:
          type: string
        period:
          type: string
        status:
          type: string
          enum: [ACTIVE, COMPLETED, ARCHIVED]
        totalEntries:
          type: integer
        highRiskCount:
          type: integer
        mediumRiskCount:
          type: integer
        lowRiskCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, clientName, period, status, createdAt, updatedAt]
    CreateEngagementBody:
      type: object
      properties:
        name:
          type: string
        clientName:
          type: string
        period:
          type: string
        description:
          type: string
      required: [name, clientName, period]

    # Upload
...
    JournalEntry:
      type: object
      properties:
        id:
          type: integer
        engagementId:
          type: integer
        entryDate:
          type: string
        postedBy:
          type: string
        description:
          type: string
        debitAccount:
          type: string
        creditAccount:
          type: string
        amount:
          type: number
        postingTime:
          type: string
        referenceNumber:
          type: string
        riskScore:
          $ref: "#/components/schemas/RiskScore"
      required: [id, engagementId, entryDate, postedBy, description, amount]
...
    RiskScore:
      type: object
      properties:
        id:
          type: integer
        entryId:
          type: integer
        totalScore:
          type: number
        riskLevel:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        postingTimeScore:
          type: number
        amountScore:
          type: number
        userConcentrationScore:
          type: number
        keywordScore:
          type: number
        frequencyScore:
          type: number
        confidenceScore:
          type: number
        overridden:
          type: boolean
        overrideReason:
          type: string
        overriddenBy:
          type: string
        overriddenAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, entryId, totalScore, riskLevel, confidenceScore, overridden]
...
    DashboardSummary:
      type: object
      properties:
        engagementId:
          type: integer
        totalEntries:
          type: integer
        highRiskCount:
          type: integer
        mediumRiskCount:
          type: integer
        lowRiskCount:
          type: integer
        highRiskPct:
          type: number
        mediumRiskPct:
          type: number
        lowRiskPct:
          type: number
        avgRiskScore:
          type: number
        topRiskyUsers:
          type: array
          items:
            $ref: "#/components/schemas/UserRiskSummary"
        recentHighRisk:
          type: array
          items:
            $ref: "#/components/schemas/JournalEntry"
        afterHoursCount:
          type: integer
        weekendCount:
          type: integer
        aiExplanationCount:
          type: integer
        duplicateSuspects:
          type: integer
      required: [engagementId, totalEntries, highRiskCount, mediumRiskCount, lowRiskCount, avgRiskScore]
    UserRiskSummary:
      type: object
      properties:
        user:
          type: string
        entryCount:
          type: integer
        highRiskCount:
          type: integer
        avgScore:
          type: number
        totalAmount:
          type: number
      required: [user, entryCount, highRiskCount, avgScore, totalAmount]
    UserHeatmapRow:
      type: object
      properties:
        user:
          type: string
        totalEntries:
          type: integer
        highRiskEntries:
          type: integer
        avgRiskScore:
          type: number
        totalAmount:
          type: number
        riskConcentration:
          type: number
      required: [user, totalEntries, highRiskEntries, avgRiskScore, totalAmount, riskConcentration]
    TimeHeatmapRow:
      type: object
      properties:
        hour:
          type: integer
        dayOfWeek:
          type: integer
        entryCount:
          type: integer
        avgRiskScore:
          type: number
        label:
          type: string
      required: [hour, dayOfWeek, entryCount, avgRiskScore]
    RiskDistribution:
      type: object
      properties:
        high:
          type: integer
        medium:
          type: integer
        low:
          type: integer
        scoreBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/ScoreRange"
      required: [high, medium, low, scoreBreakdown]
    ScoreRange:
      type: object
      properties:
        range:
          type: string
        count:
          type: integer
      required: [range, count]
    BenfordAnalysis:
      type: object
      properties:
        engagementId:
          type: integer
        digits:
          type: array
          items:
            $ref: "#/components/schemas/BenfordDigit"
        deviationScore:
          type: number
        riskAssessment:
          type: string
        summary:
          type: string
      required: [engagementId, digits, deviationScore, riskAssessment, summary]
    BenfordDigit:
      type: object
      properties:
        digit:
          type: integer
        expected:
          type: number
        actual:
          type: number
        count:
          type: integer
        deviation:
          type: number
      required: [digit, expected, actual, count, deviation]
    DuplicateGroup:
      type: object
      properties:
        groupId:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/JournalEntry"
        matchReason:
          type: string
        riskScore:
          type: number
      required: [groupId, entries, matchReason, riskScore]
...